Update Android Arthur to V10

+++ Commands/purge.js
@@ -6,10 +6,14 @@ module.exports = {
     name: "purge",
     description: "",
 	async execute(msg, args, Bot, Color, Version, Prefix) {
+        if (!global.isListening) return;
         if (typeof parseInt(args) == "number") {
             if (msg.member.hasPermission("MANAGE_MESSAGES")) {
                 msg.channel.bulkDelete(clamp(parseInt(args) + 1, 0, 100));
             } else msg.channel.send(`You do not have permission to Delete Messages`);
         }
+    },
+    init(Bot, Color, Version) {
+
     }
 };
